+++ app/src/main/java/com/rusmyhal/gitbook/ui/search/UsersAdapter.kt
@@ -1,12 +1,12 @@
 package com.rusmyhal.gitbook.ui.search
 
-import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.recyclerview.widget.DiffUtil
 import androidx.recyclerview.widget.RecyclerView
 import com.rusmyhal.gitbook.R
 import com.rusmyhal.gitbook.model.entity.User
+import com.rusmyhal.gitbook.utils.extensions.inflate
 import com.squareup.picasso.Picasso
 import kotlinx.android.synthetic.main.item_user.view.*
 
@@ -18,13 +18,7 @@ class UsersAdapter(
     private val users = arrayListOf<User>()
 
     override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): UserViewHolder {
-        return UserViewHolder(
-            LayoutInflater.from(parent.context).inflate(
-                R.layout.item_user,
-                parent,
-                false
-            )
-        )
+        return UserViewHolder(parent.inflate(R.layout.item_user))
     }
 
     override fun onBindViewHolder(holder: UserViewHolder, position: Int) {
