+++ app/src/main/java/com/rusmyhal/gitbook/utils/extensions/ViewExtensions.kt
@@ -1,0 +1,11 @@
+package com.rusmyhal.gitbook.utils.extensions
+
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import androidx.annotation.LayoutRes
+
+
+fun ViewGroup.inflate(@LayoutRes layoutRes: Int, attachToRoot: Boolean = false): View {
+    return LayoutInflater.from(context).inflate(layoutRes, this, attachToRoot)
+}
